BUG-253: Add component for showing user lists in groups
